requirements.txt
google-api-python-client==1.6.2
oauth2client==4.0.0
httplib2==0.10.3
gcloud==0.18.3